scripts: Add util.py
This commit adds common util functions for scripts.
