+++ scripts/util.py
@@ -1,0 +1,51 @@
+from subprocess import check_call
+from sys import executable
+
+
+KB = 1024
+MB = 1024 * 1024
+GB = 1024 * 1024 * 1024
+
+RED = "\033[91m"
+YELLOW = "\033[93m"
+GREEN = "\033[92m"
+RESET = "\033[0m"
+
+
+def check_and_install_zipnn():
+    try:
+        import zipnn
+    except ImportError:
+        print("zipnn not found. Installing...")
+        check_call(
+            [
+                executable,
+                "-m",
+                "pip",
+                "install",
+                "zipnn",
+                "--upgrade",
+            ]
+        )
+        import zipnn
+
+
+def parse_streaming_chunk_size(
+    streaming_chunk_size,
+):
+    if str(streaming_chunk_size).isdigit():
+        final = int(streaming_chunk_size)
+    else:
+        size_value = int(streaming_chunk_size[:-2])
+        size_unit = streaming_chunk_size[-2].lower()
+
+        if size_unit == "k":
+            final = KB * size_value
+        elif size_unit == "m":
+            final = MB * size_value
+        elif size_unit == "g":
+            final = GB * size_value
+        else:
+            raise ValueError(f"Invalid size unit: {size_unit}. Use 'k', 'm', or 'g'.")
+
+    return final
