+++ scripts/zipnn_compress_path.py
@@ -1,14 +1,14 @@
 import os
-import subprocess
 import sys
 import argparse
-from pathlib import Path
 from concurrent.futures import (
     ProcessPoolExecutor,
     as_completed,
 )
 from zipnn_compress_file import compress_file
 import multiprocessing
+from util import check_and_install_zipnn, parse_streaming_chunk_size, RED, RESET, GREEN, YELLOW
+
 sys.path.append(
     os.path.abspath(
         os.path.join(
@@ -17,60 +17,6 @@ sys.path.append(
     )
 )
 
-
-KB = 1024
-MB = 1024 * 1024
-GB = 1024 * 1024 * 1024
-
-RED = "\033[91m"
-YELLOW = "\033[93m"
-GREEN = "\033[92m"
-RESET = "\033[0m"
-
-
-def check_and_install_zipnn():
-    try:
-        import zipnn
-    except ImportError:
-        print("zipnn not found. Installing...")
-        subprocess.check_call(
-            [
-                sys.executable,
-                "-m",
-                "pip",
-                "install",
-                "zipnn",
-                "--upgrade",
-            ]
-        )
-        import zipnn
-
-
-def parse_streaming_chunk_size(
-    streaming_chunk_size,
-):
-    if str(streaming_chunk_size).isdigit():
-        final = int(streaming_chunk_size)
-    else:
-        size_value = int(
-            streaming_chunk_size[:-2]
-        )
-        size_unit = streaming_chunk_size[
-            -2
-        ].lower()
-
-        if size_unit == "k":
-            final = KB * size_value
-        elif size_unit == "m":
-            final = MB * size_value
-        elif size_unit == "g":
-            final = GB * size_value
-        else:
-            raise ValueError(
-                f"Invalid size unit: {size_unit}. Use 'k', 'm', or 'g'."
-            )
-
-    return final
 
 def replace_in_file(file_path, old: str, new: str) -> None:
     """Given a file_path, replace all occurrences of `old` with `new` inpalce."""
